Deep Equilibrium Models

Standard Models

FastDEQ.DeepEquilibriumNetwork — Type

DeepEquilibriumNetwork(model, solver; jacobian_regularization::Bool=false,
                       p=nothing, sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10),
                       kwargs...)

Deep Equilibrium Network as proposed in Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019)

Arguments

model: Explicit Neural Network which takes 2 inputs

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution

p: Optional parameters for the model

sensealg: See SteadyStateAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

model = DeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

model(rand(Float32, 2, 1))

See also: SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

FastDEQ.SkipDeepEquilibriumNetwork — Type

SkipDeepEquilibriumNetwork(model, shortcut, solver; p=nothing, jacobian_regularization::Bool=false,
                           sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10), kwargs...)
SkipDeepEquilibriumNetwork(model, solver; p=nothing, jacobian_regularization::Bool=false,
                           sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Skip Deep Equilibrium Network as proposed in Avik Pal, Alan Edelman, Christopher Rackauckas (2022)

Arguments

model: Explicit Neural Network which takes 2 inputs

shortcut: Shortcut for the network (If not given, then we create SkipDEQV2)

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution

p: Optional parameters for the model

sensealg: See SteadyStateAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

# SkipDEQ
model = SkipDeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    Dense(2, 2),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

model(rand(Float32, 2, 1))

# SkipDEQV2
model = SkipDeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

model(rand(Float32, 2, 1))

See also: DeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

MultiScale Models

FastDEQ.MultiScaleDeepEquilibriumNetwork — Type

MultiScaleDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix, solver;
                                 post_fuse_layers::Union{Tuple,Nothing}=nothing, p=nothing,
                                 sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Multiscale Deep Equilibrium Network as proposed in Shaojie Bai, Vladlen Koltun, J. Zico Kolter (2020)

Arguments

main_layers: Tuple of Explicit Neural Networks. The first network needs to take 2 inputs, the other ones only take 1 input

mapping_layers: Matrix of Explicit Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

post_fuse_layers: Tuple of Explicit Neural Networks. Applied after the mapping_layers (Default: nothing)

p: Optional parameters for the model

sensealg: See SteadyStateAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

model = MultiScaleDeepEquilibriumNetwork(
    (
        Parallel(+, Dense(4, 4, tanh_fast), Dense(4, 4, tanh_fast)),
        Dense(3, 3, tanh_fast), Dense(2, 2, tanh_fast),
        Dense(1, 1, tanh_fast)
    ),
    [
        NoOpLayer() Dense(4, 3, tanh_fast) Dense(4, 2, tanh_fast) Dense(4, 1, tanh_fast);
        Dense(3, 4, tanh_fast) NoOpLayer() Dense(3, 2, tanh_fast) Dense(3, 1, tanh_fast);
        Dense(2, 4, tanh_fast) Dense(2, 3, tanh_fast) NoOpLayer() Dense(2, 1, tanh_fast);
        Dense(1, 4, tanh_fast) Dense(1, 3, tanh_fast) Dense(1, 2, tanh_fast) NoOpLayer()
    ],
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0),
)

model(rand(Float32, 4, 1))

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

FastDEQ.MultiScaleSkipDeepEquilibriumNetwork — Type

MultiScaleSkipDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix, shortcut_layers::Tuple,
                                     solver; post_fuse_layers::Union{Tuple,Nothing}=nothing, p=nothing,
                                     sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10), kwargs...)
MultiScaleSkipDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix, solver;
                                     post_fuse_layers::Union{Tuple,Nothing}=nothing, p=nothing,
                                     sensealg=SteadyStateAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Multiscale Deep Equilibrium Network as proposed in Shaojie Bai, Vladlen Koltun, J. Zico Kolter (2020)

Arguments

main_layers: Tuple of Explicit Neural Networks. The first network needs to take 2 inputs, the other ones only take 1 input

mapping_layers: Matrix of Explicit Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer

shortcut_layers: Shortcuts for the network (If not given, then we create SkipDEQV2)

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

post_fuse_layers: Tuple of Explicit Neural Networks. Applied after the mapping_layers (Default: nothing)

p: Optional parameters for the model

sensealg: See SteadyStateAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

# MSkipDEQ
model = MultiScaleSkipDeepEquilibriumNetwork(
    (
        Parallel(+, Dense(4, 4, tanh_fast), Dense(4, 4, tanh_fast)),
        Dense(3, 3, tanh_fast), Dense(2, 2, tanh_fast),
        Dense(1, 1, tanh_fast)
    ),
    [
        NoOpLayer() Dense(4, 3, tanh_fast) Dense(4, 2, tanh_fast) Dense(4, 1, tanh_fast);
        Dense(3, 4, tanh_fast) NoOpLayer() Dense(3, 2, tanh_fast) Dense(3, 1, tanh_fast);
        Dense(2, 4, tanh_fast) Dense(2, 3, tanh_fast) NoOpLayer() Dense(2, 1, tanh_fast);
        Dense(1, 4, tanh_fast) Dense(1, 3, tanh_fast) Dense(1, 2, tanh_fast) NoOpLayer()
    ],
    (
        Dense(4, 4, tanh_fast),
        Dense(4, 3, tanh_fast),
        Dense(4, 2, tanh_fast),
        Dense(4, 1, tanh_fast)
    ),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0),
)

model(rand(Float32, 4, 1))

# MSkipDEQV2
model = MultiScaleSkipDeepEquilibriumNetwork(
    (
        Parallel(+, Dense(4, 4, tanh_fast), Dense(4, 4, tanh_fast)),
        Dense(3, 3, tanh_fast), Dense(2, 2, tanh_fast),
        Dense(1, 1, tanh_fast)
    ),
    [
        NoOpLayer() Dense(4, 3, tanh_fast) Dense(4, 2, tanh_fast) Dense(4, 1, tanh_fast);
        Dense(3, 4, tanh_fast) NoOpLayer() Dense(3, 2, tanh_fast) Dense(3, 1, tanh_fast);
        Dense(2, 4, tanh_fast) Dense(2, 3, tanh_fast) NoOpLayer() Dense(2, 1, tanh_fast);
        Dense(1, 4, tanh_fast) Dense(1, 3, tanh_fast) Dense(1, 2, tanh_fast) NoOpLayer()
    ],
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0),
)

model(rand(Float32, 4, 1))

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork