Layers

FastDEQ.DEQChain — Type

DEQChain(pre_deq, deq, post_deq)
DEQChain(layers...)

A Sequential Model containing a DEQ.

Note

The Model should contain exactly 1 AbstractDEQ Layer

FastDEQ.MultiParallelNet — Type

MultiParallelNet(layers...)
MultiParallelNet(layers::Tuple)
MultiParallelNet(layers::Vector)

Creates a MultiParallelNet mostly used for MultiScale Models. It takes a list of inputs and passes all of them through each layer and returns a tuple of outputs.

Example

Model := MultiParallelNet(L1, L2, L3)

Model(X1, X2) := (Model.L1(X1, X2), Model.L2(X1, X2), Model.L3(X1, X2))