Miscellaneous

DiffEqSensitivity.SteadyStateAdjoint — Type

SteadyStateAdjoint(reltol, abstol, maxiters; autojacvec=ZygoteVJP(),
                   linsolve=KrylovJL_GMRES(; rtol=reltol, atol=abstol, itmax=maxiters))

Creates SteadyStateAdjoint (Steven G Johnson (2006)) with sensible defaults.

Arguments

reltol: Relative tolerance.

abstol: Absolute tolerance.

maxiters: Maximum number of iterations.

autojacvec: Which backend to use for VJP.

linsolve: Linear Solver from LinearSolve.jl.

FastDEQ.DeepEquilibriumSolution — Type

DeepEquilibriumSolution(z_star, u₀, residual, jacobian_loss, nfe)

Stores the solution of a DeepEquilibriumNetwork and its variants.

Fields

z_star: Steady-State or the value reached due to maxiters

u₀: Initial Condition

residual: Difference of the $z^*$ and $f(z^*, x)$

jacobian_loss: Jacobian Stabilization Loss (see individual networks to see how it can be computed)

nfe: Number of Function Evaluations

FastDEQ.get_and_clear_nfe! — Function

get_and_clear_nfe!(model::AbstractDeepEquilibriumNetwork)

Return the number of function evaluations (NFE) and clear the counter.

FastDEQ.compute_deq_jacobian_loss — Function

compute_deq_jacobian_loss(re, p, z, x)

Computes Jacobian Stabilization Loss (Shaojie Bai, Vladlen Koltun, J Zico Kolter (2021)).

Arguments

re: Constructs the model given the parameters p.

p: Parameters of the model.

z: Steady State.

x: Input to the model.

Current Known Failure Modes

Conv layers error out due to ForwardDiff on GPUs

If the model internally uses destructure/restructure eg. WeightNorm Layer, then this loss function will error out in the backward pass.

FastDEQ.NormalInitializer — Function

NormalInitializer(μ = 0.0f0, σ² = 0.01f0)

Initializes the weights of the network with a normal distribution. For DEQs the training is stable if we use this as the Initialization

FastDEQ.SupervisedLossContainer — Type

SupervisedLossContainer(loss_function)
SupervisedLossContainer(loss_function, λ, λⱼ)

A container class for supervised loss functions.