Dynamical System Variants

Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019) introduced Discrete Deep Equilibrium Models which drives a Discrete Dynamical System to its steady-state. Avik Pal, Alan Edelman, Christopher Rackauckas (2022) extends this framework to Continuous Dynamical Systems which converge to the steady-stable in a more stable fashion. For a detailed discussion refer to Avik Pal, Alan Edelman, Christopher Rackauckas (2022).

Continuous DEQs (Infinite Time Neural ODEs)

FastDEQ.ContinuousDEQSolver — Type

ContinuousDEQSolver(alg=VCABM4(); mode::Symbol=:rel_deq_default, abstol=1e-8, reltol=1e-8, tspan=Inf)

Solver for Continuous DEQ Problem (Avik Pal, Alan Edelman, Christopher Rackauckas (2022)). Similar to DynamicSS but provides more flexibility needed for solving DEQ problems.

Arguments

alg: Algorithm to solve the ODEProblem. (Default: VCABM4())

mode: Termination Mode of the solver. See below for a description of the various termination conditions (Default: :rel_deq_default)

abstol: Absolute tolerance for termination. (Default: 1e-8)

reltol: Relative tolerance for termination. (Default: 1e-8)

tspan: Time span. Users should not change this value, instead control termination through maxiters in solve (Default: Inf)

Termination Modes

Termination on Absolute Tolerance

:abs: Terminates if $all \left( \frac{\partial u}{\partial t} \| \leq abstol \right)$

:abs_norm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq abstol$

:abs_deq_default: Essentially abs_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

:abs_deq_best: Same as :abs_deq_default but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination on Relative Tolerance

:rel: Terminates if $all \left(\frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t}\| \right)$

:rel_norm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$

:rel_deq_best: Essentially rel_norm + terminate if there has been no improvement for the last 30 steps + terminate if the solution blows up (diverges)

:rel_deq_default: Same as :rel_deq_default but uses the best solution found so far, i.e. deviates only if the solution has not converged

Termination using both Absolute and Relative Tolerances

:norm: Terminates if $\| \frac{\partial u}{\partial t} \| \leq reltol \times \| \frac{\partial u}{\partial t} + u \|$ & $\| \frac{\partial u}{\partial t} \| \leq abstol$

fallback: Check if all values of the derivative is close to zero wrt both relative and absolute tolerance. This is usable for small problems but doesn't scale well for neural networks, and should be avoided unless absolutely necessary

See also: DiscreteDEQSolver

Note

This will be upstreamed to DiffEqSensitivity in the later releases of the package

Discrete DEQs

FastDEQ.DiscreteDEQSolver — Function

DiscreteDEQSolver(solver=LimitedMemoryBroydenSolver; abstol=1e-8, reltol=1e-8, kwargs...)

Solver for Discrete DEQ Problem (Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019)). A wrapper around SSrootfind to mimic the ContinuousDEQSolver API.

Arguments

solver: NonLinear Solver for the DEQ problem. (Default: LimitedMemoryBroydenSolver)

abstol: Absolute tolerance for termination. (Default: 1e-8)

reltol: Relative tolerance for termination. (Default: 1e-8)

kwargs: Additional keyword arguments passed to the solver.

Note

There is no mode kwarg for DiscreteDEQSolver. Instead solvers directly define their own termination condition. For BroydenSolver and LimitedMemoryBroydenSolver, the termination conditions are :abs_norm & :rel_deq_default respectively.

See also: ContinuousDEQSolver