Deep Equilibrium Models

Standard Models

DeepEquilibriumNetworks.DeepEquilibriumNetwork — Type

DeepEquilibriumNetwork(model, solver; jacobian_regularization::Bool=false,
                       sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Deep Equilibrium Network as proposed in Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019).

Arguments

model:Lux Neural Network.

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution.

sensealg: See DeepEquilibriumAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

import DeepEquilibriumNetworks as DEQs
import Lux
import Random
import OrdinaryDiffEq

model = DEQs.DeepEquilibriumNetwork(Lux.Parallel(+, Lux.Dense(2, 2; bias=false),
                                                 Lux.Dense(2, 2; bias=false)),
                                    DEQs.ContinuousDEQSolver(OrdinaryDiffEq.VCABM3();
                                                             abstol=0.01f0, reltol=0.01f0))

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

See also: SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork.

DeepEquilibriumNetworks.SkipDeepEquilibriumNetwork — Type

SkipDeepEquilibriumNetwork(model, shortcut, solver; jacobian_regularization::Bool=false,
                           sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Skip Deep Equilibrium Network as proposed in Avik Pal, Alan Edelman, Christopher Rackauckas (2022)

Arguments

model: Neural Network.

shortcut: Shortcut for the network (pass nothing for SkipDEQV2).

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution.

sensealg: See DeepEquilibriumAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

import DeepEquilibriumNetworks as DEQs
import Lux
import Random
import OrdinaryDiffEq

## SkipDEQ
model = DEQs.SkipDeepEquilibriumNetwork(Lux.Parallel(+, Lux.Dense(2, 2; bias=false),
                                                     Lux.Dense(2, 2; bias=false)),
                                        Lux.Dense(2, 2),
                                        DEQs.ContinuousDEQSolver(OrdinaryDiffEq.VCABM3();
                                                                 abstol=0.01f0,
                                                                 reltol=0.01f0))

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

## SkipDEQV2
model = DEQs.SkipDeepEquilibriumNetwork(Lux.Parallel(+, Lux.Dense(2, 2; bias=false),
                                                     Lux.Dense(2, 2; bias=false)), nothing,
                                        DEQs.ContinuousDEQSolver(OrdinaryDiffEq.VCABM3();
                                                                 abstol=0.01f0,
                                                                 reltol=0.01f0))

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

See also: DeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

MultiScale Models

Missing docstring.

Missing docstring for MultiScaleDeepEquilibriumNetwork. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MultiScaleSkipDeepEquilibriumNetwork. Check Documenter's build log for details.