Miscellaneous

DeepEquilibriumNetworks.DeepEquilibriumSolution — Type

DeepEquilibriumSolution(z_star, u₀, residual, jacobian_loss, nfe)

Stores the solution of a DeepEquilibriumNetwork and its variants.

Fields

z_star: Steady-State or the value reached due to maxiters

u0: Initial Condition

residual: Difference of the $z^*$ and $f(z^*, x)$

jacobian_loss: Jacobian Stabilization Loss (see individual networks to see how it can be computed).

nfe: Number of Function Evaluations

DeepEquilibriumNetworks.EquilibriumSolution — Type

EquilibriumSolution

Wraps the solution of a SteadyStateProblem using either ContinuousDEQSolver or DiscreteDEQSolver. This is mostly an internal implementation detail, which allows proper dispatch during adjoint computation without type piracy.

DeepEquilibriumNetworks.split_and_reshape — Function

split_and_reshape(x::AbstractMatrix, ::Val{idxs}, ::Val{shapes}) where {idxs, shapes}

Splits up the AbstractMatrix into chunks and reshapes them.

Arguments

x: Matrix to be split up.

Sidxs: Indices to partition the array at. (must be a Val type).

Sshapes: Shapes to reshape the split the arrays. (must be a Val type).

Example

using DeepEquilibriumNetworks, Static

x1 = ones(Float32, 4, 4)
x2 = fill!(zeros(Float32, 2, 4), 0.5f0)
x3 = zeros(Float32, 1, 4)

x = vcat(x1, x2, x3)
split_idxs = Val(cumsum((0, size(x1, 1), size(x2, 1), size(x3, 1))))
shapes = Val((size(x1, 1), size(x2, 1), size(x3, 1)))

DEQs.split_and_reshape(x, split_idxs, shapes)

DeepEquilibriumNetworks.estimate_jacobian_trace — Function

estimate_jacobian_trace(::Val{mode}, model::Lux.AbstractExplicitLayer, ps,
                        st::NamedTuple, z::AbstractArray, x::AbstractArray,
                        rng::Random.AbstractRNG)

Estimates the trace of the jacobian matrix wrt z.

Arguments

mode: Options: reverse and finite_diff

model: A Lux Neural Network mapping 2 equal sized arrays to a same sized array. This convention is not checked, and if violated will lead to errors.

ps: Parameters of model.

st: States of model.

z: Input wrt the Jacobian is computed.

x: Other Input to model.

rng: PRNG. Note that this object is mutated by this function.

Returns

Stochastic Estimate of the trace of the Jacobian.