NonLinear Solvers

We provide the following NonLinear Solvers for DEQs. These are compatible with GPUs.

Note

If you are looking for general purpose nonlinear solvers, we recommend checking out NonlinearSolve.jl

DeepEquilibriumNetworks.BroydenSolver — Type

BroydenSolver(; T=Float32, device, original_dims::Tuple{Int,Int}, batch_size, maxiters::Int=50, ϵ::Real=1e-6,
              abstol::Union{Real,Nothing}=nothing, reltol::Union{Real,Nothing}=nothing)

Broyden Solver (Charles G Broyden (1965)) for solving Discrete DEQs. It is recommended to use LimitedMemoryBroydenSolver for better performance.

Arguments

T: The type of the elements of the vectors. (Default: Float32)

device: The device to use. Pass gpu to use the GPU else pass cpu.

original_dims: Dimensions to reshape the arrays into (excluding the batch dimension).

batch_size: The batch size of the problem. Your inputs can have a different batch size, but having them match allows us to efficiently cache internal statistics without reallocation.

maxiters: Maximum number of iterations to run.

ϵ: Tolerance for convergence.

abstol: Absolute tolerance.

reltol: Relative tolerance. (This value is ignored by BroydenSolver at the moment)

See also: LimitedMemoryBroydenSolver

DeepEquilibriumNetworks.LimitedMemoryBroydenSolver — Type

LimitedMemoryBroydenSolver(; T=Float32, device, original_dims::Tuple{Int,Int}, batch_size, maxiters::Int=50,
                           ϵ::Real=1e-6, criteria::Symbol=:reltol, abstol::Union{Real,Nothing}=nothing,
                           reltol::Union{Real,Nothing}=nothing

Limited Memory Broyden Solver (Shaojie Bai, Vladlen Koltun, J. Zico Kolter (2020)) for solving Discrete DEQs.

Arguments

T: The type of the elements of the vectors. (Default: Float32)

device: The device to use. Pass gpu to use the GPU else pass cpu.

original_dims: Dimensions to reshape the arrays into (excluding the batch dimension).

batch_size: The batch size of the problem. Your inputs can have a different batch size, but having them match allows us to efficiently cache internal statistics without reallocation.

maxiters: Maximum number of iterations to run.

ϵ: Tolerance for convergence.

criteria: The criteria to use for convergence. Can be :reltol or :abstol.

abstol: Absolute tolerance.

reltol: Relative tolerance.

See also: BroydenSolver