Deep Equilibrium Models

Standard Models

FastDEQ.DeepEquilibriumNetwork — Type

DeepEquilibriumNetwork(model, solver; jacobian_regularization::Bool=false, sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Deep Equilibrium Network as proposed in Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019)

Arguments

model: Neural Network

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution

sensealg: See DeepEquilibriumAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

model = DeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(Float32, 2, 1), ps, st)

See also: SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

FastDEQ.SkipDeepEquilibriumNetwork — Type

SkipDeepEquilibriumNetwork(model, shortcut, solver; jacobian_regularization::Bool=false, sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Skip Deep Equilibrium Network as proposed in Avik Pal, Alan Edelman, Christopher Rackauckas (2022)

Arguments

model: Neural Network

shortcut: Shortcut for the network (pass nothing for SkipDEQV2)

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution

sensealg: See DeepEquilibriumAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

# SkipDEQ
model = SkipDeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    Dense(2, 2),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(Float32, 2, 1), ps, st)

# SkipDEQV2
model = SkipDeepEquilibriumNetwork(
    Parallel(
        +,
        Dense(2, 2; bias=false),
        Dense(2, 2; bias=false)
    ),
    nothing,
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)
)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(Float32, 2, 1), ps, st)

See also: DeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

MultiScale Models

FastDEQ.MultiScaleDeepEquilibriumNetwork — Type

MultiScaleDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix, post_fuse_layer::Union{Nothing,Tuple}, solver, scales; sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Multiscale Deep Equilibrium Network as proposed in Shaojie Bai, Vladlen Koltun, J. Zico Kolter (2020)

Arguments

main_layers: Tuple of Neural Networks. The first network needs to take a tuple of 2 arrays, the other ones only take 1 input

mapping_layers: Matrix of Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer

post_fuse_layer: Tuple of Neural Networks. Each of the scales are passed through this layer

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

scales: Output scales

sensealg: See DeepEquilibriumAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

model = MultiScaleDeepEquilibriumNetwork(
    (
        Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
        Dense(3, 3, tanh),
        Dense(2, 2, tanh),
        Dense(1, 1, tanh)
    ),
    [
        NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh);
        Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh);
        Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh);
        Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()
    ],
    nothing,
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0),
    ((4,), (3,), (2,), (1,)),
)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 1)

model(x, ps, st)

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

FastDEQ.MultiScaleSkipDeepEquilibriumNetwork — Type

MultiScaleSkipDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix, post_fuse_layer::Union{Nothing,Tuple}, shortcut_layers::Union{Nothing,Tuple}, solver, scales; sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Multiscale Deep Equilibrium Network as proposed in Shaojie Bai, Vladlen Koltun, J. Zico Kolter (2020) combined with Skip Deep Equilibrium Network as proposed in Avik Pal, Alan Edelman, Christopher Rackauckas (2022)

Arguments

main_layers: Tuple of Neural Networks. The first network needs to take a tuple of 2 arrays, the other ones only take 1 input

mapping_layers: Matrix of Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer

post_fuse_layer: Tuple of Neural Networks. Each of the scales are passed through this layer

shortcut_layers: Shortcut for the network (pass nothing for SkipDEQV2)

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver)

scales: Output scales

sensealg: See DeepEquilibriumAdjoint

kwargs: Additional Parameters that are directly passed to solve

Example

# MSkipDEQ
model = MultiScaleSkipDeepEquilibriumNetwork(
    (
        Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
        Dense(3, 3, tanh),
        Dense(2, 2, tanh),
        Dense(1, 1, tanh),
    ),
    [
        NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh)
        Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh)
        Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh)
        Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()
    ],
    nothing,
    (Dense(4, 4, tanh), Dense(4, 3, tanh), Dense(4, 2, tanh), Dense(4, 1, tanh)),
    ContinuousDEQSolver(; abstol=0.1f0, reltol=0.1f0, abstol_termination=0.1f0, reltol_termination=0.1f0),
    ((4,), (3,), (2,), (1,));
    sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10),
)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 2)

model(x, ps, st)

# MSkipDEQV2
model = MultiScaleSkipDeepEquilibriumNetwork(
        (
            Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
            Dense(3, 3, tanh),
            Dense(2, 2, tanh),
            Dense(1, 1, tanh),
        ),
        [
            NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh)
            Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh)
            Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh)
            Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()
        ],
        nothing,
        nothing,
        ContinuousDEQSolver(; abstol=0.1f0, reltol=0.1f0, abstol_termination=0.1f0, reltol_termination=0.1f0),
        ((4,), (3,), (2,), (1,));
        sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10),
    )

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 2)

model(x, ps, st)

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork