Deep Equilibrium Models

Standard Models

DeepEquilibriumNetworks.DeepEquilibriumNetwork — Type

DeepEquilibriumNetwork(model, solver; jacobian_regularization::Bool=false,
                       sensealg=SteadyStateAdjoint(), kwargs...)

Deep Equilibrium Network as proposed in (Bai *et al.*, 2019).

Arguments

model:Lux Neural Network.

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution.

sensealg: See SciMLSensitivity.SteadyStateAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

using DeepEquilibriumNetworks, Lux, Random, OrdinaryDiffEq

model = DeepEquilibriumNetwork(Parallel(+,
        Dense(2, 2; use_bias=false),
        Dense(2, 2; use_bias=false)),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0); save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

See also: SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork.

DeepEquilibriumNetworks.SkipDeepEquilibriumNetwork — Type

SkipDeepEquilibriumNetwork(model, shortcut, solver; jacobian_regularization::Bool=false,
                           sensealg=DeepEquilibriumAdjoint(0.1f0, 0.1f0, 10), kwargs...)

Skip Deep Equilibrium Network as proposed in (Pal *et al.*, 2022)

Arguments

model: Neural Network.

shortcut: Shortcut for the network (pass nothing for SkipRegDEQ).

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

jacobian_regularization: If true, Jacobian Loss is computed and stored in the DeepEquilibriumSolution.

sensealg: See SciMLSensitivity.SteadyStateAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

using DeepEquilibriumNetworks, Lux, Random, OrdinaryDiffEq

## SkipDEQ
model = SkipDeepEquilibriumNetwork(Parallel(+,
        Dense(2, 2; use_bias=false),
        Dense(2, 2; use_bias=false)),
    Dense(2, 2),
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0); save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

## SkipRegDEQ
model = SkipDeepEquilibriumNetwork(Parallel(+,
        Dense(2, 2; use_bias=false),
        Dense(2, 2; use_bias=false)),
    nothing,
    ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0); save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)

model(rand(rng, Float32, 2, 1), ps, st)

See also: DeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

MultiScale Models

DeepEquilibriumNetworks.MultiScaleDeepEquilibriumNetwork — Type

MultiScaleDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix,
                                 post_fuse_layer::Union{Nothing,Tuple}, solver, scales;
                                 sensealg=SteadyStateAdjoint(), kwargs...)

Multiscale Deep Equilibrium Network as proposed in (Bai *et al.*, 2020)

Arguments

main_layers: Tuple of Neural Networks. The first network needs to take a tuple of 2 arrays, the other ones only take 1 input.

mapping_layers: Matrix of Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer.

post_fuse_layer: Tuple of Neural Networks. Each of the scales is passed through this layer.

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

scales: Output scales.

sensealg: See SciMLSensitivity.SteadyStateAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

using DeepEquilibriumNetworks, Lux, Random, OrdinaryDiffEq

main_layers = (Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
    Dense(3, 3, tanh),
    Dense(2, 2, tanh),
    Dense(1, 1, tanh))

mapping_layers = [NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh);
    Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh);
    Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh);
    Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()]

solver = ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)

model = MultiScaleDeepEquilibriumNetwork(main_layers, mapping_layers, nothing,
    solver, ((4,), (3,), (2,), (1,)); save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 1)

model(x, ps, st)

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork

DeepEquilibriumNetworks.MultiScaleSkipDeepEquilibriumNetwork — Type

MultiScaleSkipDeepEquilibriumNetwork(main_layers::Tuple, mapping_layers::Matrix,
                                     post_fuse_layer::Union{Nothing,Tuple},
                                     shortcut_layers::Union{Nothing,Tuple}, solver,
                                     scales; sensealg=SteadyStateAdjoint(), kwargs...)

Multiscale Deep Equilibrium Network as proposed in (Bai *et al.*, 2020) combined with Skip Deep Equilibrium Network as proposed in (Pal *et al.*, 2022).

Arguments

main_layers: Tuple of Neural Networks. The first network needs to take a tuple of 2 arrays, the other ones only take 1 input.

mapping_layers: Matrix of Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer.

post_fuse_layer: Tuple of Neural Networks. Each of the scales is passed through this layer.

shortcut_layers: Shortcut for the network (pass nothing for SkipRegDEQ).

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

scales: Output scales.

sensealg: See SciMLSensitivity.SteadyStateAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

using DeepEquilibriumNetworks, Lux, Random, OrdinaryDiffEq

# MSkipDEQ
main_layers = (Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
    Dense(3, 3, tanh),
    Dense(2, 2, tanh),
    Dense(1, 1, tanh))

mapping_layers = [NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh);
    Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh);
    Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh);
    Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()]

solver = ContinuousDEQSolver(VCABM3(); abstol=0.01f0, reltol=0.01f0)

shortcut_layers = (Dense(4, 4, tanh),
    Dense(4, 3, tanh),
    Dense(4, 2, tanh),
    Dense(4, 1, tanh))

model = MultiScaleSkipDeepEquilibriumNetwork(main_layers,
    mapping_layers,
    nothing,
    shortcut_layers,
    solver,
    ((4,), (3,), (2,), (1,));
    save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 2)

model(x, ps, st)

# MSkipRegDEQ
model = MultiScaleSkipDeepEquilibriumNetwork(main_layers,
    mapping_layers,
    nothing,
    nothing,
    solver,
    ((4,), (3,), (2,), (1,));
    save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 2)

model(x, ps, st)

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork

DeepEquilibriumNetworks.MultiScaleNeuralODE — Type

MultiScaleNeuralODE(main_layers::Tuple, mapping_layers::Matrix,
                    post_fuse_layer::Union{Nothing,Tuple}, solver, scales;
                    sensealg=GaussAdjoint(; autojacvec=ZygoteVJP()), kwargs...)

Multiscale Neural ODE with Input Injection.

Arguments

main_layers: Tuple of Neural Networks. The first network needs to take a tuple of 2 arrays, the other ones only take 1 input.

mapping_layers: Matrix of Neural Networks. The $(i, j)^{th}$ network takes the output of $i^{th}$ main_layer and passes it to the $j^{th}$ main_layer.

post_fuse_layer: Tuple of Neural Networks. Each of the scales is passed through this layer.

solver: Solver for the optimization problem (See: ContinuousDEQSolver & DiscreteDEQSolver).

scales: Output scales.

sensealg: See SciMLSensitivity.InterpolatingAdjoint.

kwargs: Additional Parameters that are directly passed to SciMLBase.solve.

Example

using DeepEquilibriumNetworks, Lux, Random, OrdinaryDiffEq

main_layers = (Parallel(+, Dense(4, 4, tanh), Dense(4, 4, tanh)),
    Dense(3, 3, tanh),
    Dense(2, 2, tanh),
    Dense(1, 1, tanh))

mapping_layers = [NoOpLayer() Dense(4, 3, tanh) Dense(4, 2, tanh) Dense(4, 1, tanh);
    Dense(3, 4, tanh) NoOpLayer() Dense(3, 2, tanh) Dense(3, 1, tanh);
    Dense(2, 4, tanh) Dense(2, 3, tanh) NoOpLayer() Dense(2, 1, tanh);
    Dense(1, 4, tanh) Dense(1, 3, tanh) Dense(1, 2, tanh) NoOpLayer()]

model = MultiScaleNeuralODE(main_layers,
    mapping_layers,
    nothing,
    VCAB3(),
    ((4,), (3,), (2,), (1,));
    save_everystep=true)

rng = Random.default_rng()
ps, st = Lux.setup(rng, model)
x = rand(rng, Float32, 4, 1)

model(x, ps, st)

See also: DeepEquilibriumNetwork, SkipDeepEquilibriumNetwork, MultiScaleDeepEquilibriumNetwork, MultiScaleSkipDeepEquilibriumNetwork