Miscellaneous

DeepEquilibriumNetworks.DeepEquilibriumAdjoint — Type

DeepEquilibriumAdjoint(reltol, abstol, maxiters; autojacvec=ZygoteVJP(),
                       linsolve=KrylovJL_GMRES(; rtol=reltol, atol=abstol, itmax=maxiters),
                       mode=:vanilla)

Creates DeepEquilibriumAdjoint (Steven G Johnson (2006)) with sensible defaults.

Arguments

reltol: Relative tolerance.

abstol: Absolute tolerance.

maxiters: Maximum number of iterations.

autojacvec: Which backend to use for VJP.

linsolve: Linear Solver from LinearSolve.jl.

mode: Adjoint mode. Currently only :vanilla & :jfb are supported.

DeepEquilibriumNetworks.DeepEquilibriumSolution — Type

DeepEquilibriumSolution(z_star, u₀, residual, jacobian_loss, nfe)

Stores the solution of a DeepEquilibriumNetwork and its variants.

Fields

* `z_star`: Steady-State or the value reached due to maxiters
* `u₀`: Initial Condition
* `residual`: Difference of the ``z^*`` and ``f(z^*, x)``
* `jacobian_loss`: Jacobian Stabilization Loss (see individual networks to see how it can be computed)
* `nfe`: Number of Function Evaluations

DeepEquilibriumNetworks.NormalInitializer — Function

NormalInitializer(μ = 0.0f0, σ² = 0.01f0)

Initializes the weights of the network with a normal distribution. For DEQs the training is stable if we use this as the Initialization