Dynamical System Variants

Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019) introduced Discrete Deep Equilibrium Models which drives a Discrete Dynamical System to its steady-state. Avik Pal, Alan Edelman, Christopher Rackauckas (2022) extends this framework to Continuous Dynamical Systems which converge to the steady-stable in a more stable fashion. For a detailed discussion refer to Avik Pal, Alan Edelman, Christopher Rackauckas (2022).

Continuous DEQs

DeepEquilibriumNetworks.ContinuousDEQSolver — Type

ContinuousDEQSolver(alg=VCABM3(); mode=SteadyStateTerminationMode.RelSafeBest,
                    abstol=1.0f-8, reltol=1.0f-6, abstol_termination=abstol,
                    reltol_termination=reltol, tspan=Inf32, kwargs...)

Solver for Continuous DEQ Problem (Avik Pal, Alan Edelman, Christopher Rackauckas (2022)). Effectively a wrapper around DynamicSS with more sensible defaults for DEQs.

Arguments

alg: Algorithm to solve the ODEProblem. (Default: VCAB3())

mode: Termination Mode of the solver. See the documentation for NLSolveTerminationCondition for more information. (Default: NLSolveTerminationMode.RelSafeBest)

abstol: Absolute tolerance for time stepping. (Default: 1f-8)

reltol: Relative tolerance for time stepping. (Default: 1f-6)

abstol_termination: Absolute tolerance for termination. (Default: abstol)

reltol_termination: Relative tolerance for termination. (Default: reltol)

tspan: Time span. Users should not change this value, instead control termination through maxiters in solve. (Default: Inf32)

kwargs: Additional Parameters that are directly passed to NLSolveTerminationCondition.

See also: DiscreteDEQSolver

Discrete DEQs

DeepEquilibriumNetworks.DiscreteDEQSolver — Type

DiscreteDEQSolver(alg = LBroyden(; batched=true,
                                 termination_condition=NLSolveTerminationCondition(NLSolveTerminationMode.RelSafe;
                                                                                   abstol=1.0f-8, reltol=1.0f-6))

Solver for Discrete DEQ Problem (Shaojie Bai, J. Zico Kolter, Vladlen Koltun (2019)). Similar to SSrootfind but provides more flexibility needed for solving DEQ problems.

Arguments

alg: Algorithm to solve the Nonlinear Problem. (Default: LBroyden)

See also: ContinuousDEQSolver